REF: move functions out of _testing/__init__ (#38961)
